refactor: 커스텀 hook으로 로직 분리
